+++ esp32-hal/examples/embassy_wait.rs
@@ -6,12 +6,11 @@
 #![no_main]
 #![feature(type_alias_impl_trait)]
 
-use embassy_executor::Executor;
 use embassy_time::{Duration, Timer};
 use embedded_hal_async::digital::Wait;
 use esp32_hal::{
     clock::ClockControl,
-    embassy,
+    embassy::{self, executor::Executor},
     gpio::{Gpio0, Input, PullDown},
     peripherals::Peripherals,
     prelude::*,
@@ -20,7 +19,7 @@ use esp32_hal::{
     IO,
 };
 use esp_backtrace as _;
-use static_cell::StaticCell;
+use static_cell::make_static;
 
 #[embassy_executor::task]
 async fn ping(mut pin: Gpio0<Input<PullDown>>) {
@@ -31,8 +30,6 @@ async fn ping(mut pin: Gpio0<Input<PullDown>>) {
         Timer::after(Duration::from_millis(100)).await;
     }
 }
-
-static EXECUTOR: StaticCell<Executor> = StaticCell::new();
 
 #[entry]
 fn main() -> ! {
@@ -74,7 +71,7 @@ fn main() -> ! {
     )
     .unwrap();
 
-    let executor = EXECUTOR.init(Executor::new());
+    let executor = make_static!(Executor::new());
     executor.run(|spawner| {
         spawner.spawn(ping(input)).ok();
     });