+++ esp32c3-hal/examples/embassy_wait.rs
@@ -20,7 +20,7 @@ use esp32c3_hal::{
     IO,
 };
 use esp_backtrace as _;
-use static_cell::StaticCell;
+use static_cell::make_static;
 
 #[embassy_executor::task]
 async fn ping(mut pin: Gpio9<Input<PullDown>>) {
@@ -31,8 +31,6 @@ async fn ping(mut pin: Gpio9<Input<PullDown>>) {
         Timer::after(Duration::from_millis(100)).await;
     }
 }
-
-static EXECUTOR: StaticCell<Executor> = StaticCell::new();
 
 #[entry]
 fn main() -> ! {
@@ -81,7 +79,7 @@ fn main() -> ! {
     )
     .unwrap();
 
-    let executor = EXECUTOR.init(Executor::new());
+    let executor = make_static!(Executor::new());
     executor.run(|spawner| {
         spawner.spawn(ping(input)).ok();
     });