+++ esp32s3-hal/examples/embassy_i2c.rs
@@ -14,11 +14,10 @@
 #![no_main]
 #![feature(type_alias_impl_trait)]
 
-use embassy_executor::Executor;
 use embassy_time::{Duration, Timer};
 use esp32s3_hal::{
     clock::ClockControl,
-    embassy,
+    embassy::{self, executor::Executor},
     i2c::I2C,
     interrupt,
     peripherals::{Interrupt, Peripherals, I2C0},
@@ -29,7 +28,7 @@ use esp32s3_hal::{
 };
 use esp_backtrace as _;
 use lis3dh_async::{Lis3dh, Range, SlaveAddr};
-use static_cell::StaticCell;
+use static_cell::make_static;
 
 #[embassy_executor::task]
 async fn run(i2c: I2C<'static, I2C0>) {
@@ -43,8 +42,6 @@ async fn run(i2c: I2C<'static, I2C0>) {
         Timer::after(Duration::from_millis(100)).await;
     }
 }
-
-static EXECUTOR: StaticCell<Executor> = StaticCell::new();
 
 #[entry]
 fn main() -> ! {
@@ -86,7 +83,7 @@ fn main() -> ! {
 
     interrupt::enable(Interrupt::I2C_EXT0, interrupt::Priority::Priority1).unwrap();
 
-    let executor = EXECUTOR.init(Executor::new());
+    let executor = make_static!(Executor::new());
     executor.run(|spawner| {
         spawner.spawn(run(i2c0)).ok();
     });